requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-admin-persian-fonts==0.2
django-filter==23.2
django-rosetta==0.9.9
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
ghasedakpack==0.1.13
idna==3.4
Markdown==3.4.3
Pillow==9.5.0
polib==1.2.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
